+++ packages/engine/src/index.ts
@@ -9,6 +9,7 @@ export * from "./models/subModel";
 export * from "./models/boxedScale";
 export * from "./models/units";
 export * from "./models/channelOutput";
+export * from "./models/generateCustom";
 export * from "./models/matrix";
 export * from "./models/line";
 export * from "./models/arches";